Add more comprehensive checking.
With this change, the script now also checks for the presence of
repository metadata for the repository that would be used to install
Netdata. If the other checks pass but this one fails, a different error
message is show indicating that packages aren’t available yet, but will
be soon, with the stable install message providing info on how to
install anyway.