Switch to XZ compression by default for DEB packages.
XZ compression is the de-facto standard compression used for DEB
packages on Ubuntu prior to Ubuntu 21.10 and Debian prior to Debian 12.

Using XZ for DEB package compression incurs a nontrivial performance
hit for the installation of the packages (specifically, the process of
unpacking the packages), but provides a roughly 34% space savings for
the whole collection of packages compared to our current usage of gzip
compression (tested using Debian 13 packages) and unlike the newer zstd
compression has been supported by dpkg for long enough that we don’t
really have to worry about any given system not supporting it.