Enable zstd compression for DEB packages on platforms that support it.
Zstd compression is supported, and used by default, in Debian and Ubuntu
starting with Debian 12 and Ubuntu 21.10.

Compared to gzip compression, zstd provides a roughly 8.6% reduction in
the total size of a full set of our packages (tested using Debian 13
packages), while at the same time reducing the overhead of unpacking the
package for installation.

Compared to the XZ compression that the previous commit enabled by
default for our packages, zstd results in roughly 38% larger packages,
but it completely removes the overhead on installation that XZ imposes,
and is thus generally considered enough of a net improvement that both
Debian and Ubuntu have pivoted from using XZ for package compression for
their distribution packages to using zstd.