Detect in kickstart script if we publish native packages for the platform (#21262)
* Detect in kickstart script if we publish native packages for the platform.

This allows more intelligent fallback behavior, insisting on using
native packages if they are being published (and are actually available)
and only falling back if they aren’t. It also means that users
installing on platforms we still have packages available for but no
longer publish packages for will end up with static installs instead of
native package installs.

This change includes a new option for the kickstart script called
`--install-type`, which can be used to specify the install type to use.
It supports five possible values currently, `native`, `static`, `build`,
`auto`, and `any`. The first three are equivalent to the existing
`--native-only`, `--static-only`, and `--build-only` options. The `any`
value indicates to use our existing fallback behavior (try each install
type in order until we find one that works). The `auto` value, which is
the new default, acts like `native` if the script detects that we
publish native packages for the platform, and `any` otherwise.

This is dependent on a flag file on the repository server that is
generated by our package upload processing code for platforms that we
publish native packages for.

* Add more comprehensive checking.

With this change, the script now also checks for the presence of
repository metadata for the repository that would be used to install
Netdata. If the other checks pass but this one fails, a different error
message is show indicating that packages aren’t available yet, but will
be soon, with the stable install message providing info on how to
install anyway.

* Actually add --install-type option.

* Improve error message.

* Fix early bail-out cases for native package installs.

They need to defer instead of failing outright when running in auto
mode.

Also, fix the exit code for `--repositories-only` installs.

* Fix incorrect error message.

* Fix error messages.