Don’t duplicate fetching of installed agent version in updater.
In the native package handling code, instead of invoking `netdata -V` a
second time, just use the parsed version number to determine if
something is a nightly build or a stable build.

This matches on whether the final five digits of the prased version
number are all zeroes (which indicates a stable build) or not.